Fix Unit Test

+++ src/modules/todo/service/todo.service.ts
@@ -1,3 +1,4 @@
+import { ValidationError } from '../../../utils/errors/index';
 import { TodoRepository } from './../repository/todo.repository';
 
 export class TodoService {
@@ -10,7 +11,18 @@ export class TodoService {
   }
 
   public async getOne(id: string) {
-    return await this.todoRepository.findOne(id);
+    const exist = await this.todoRepository.findOne(id);
+    if(!exist) {
+      throw {
+        value: {
+          code: 404,
+          message: 'Not Found'
+        },
+        error: new ValidationError('Not Found')
+      }
+    }
+  
+    return exist;
   }
 
   public async create(val: object) {
@@ -18,10 +30,32 @@ export class TodoService {
   }
 
   public async update(id: string, val: object) {
+    const exist = await this.todoRepository.findOne(id);
+    if(!exist) {
+      throw {
+        value: {
+          code: 404,
+          message: 'Not Found'
+        },
+        error: new ValidationError('Not Found')
+      }
+    }
+
     return await this.todoRepository.update(id, val);
   }
 
   public async delete(id: string) {
+    const exist = await this.todoRepository.findOne(id);
+    if(!exist) {
+      throw {
+        value: {
+          code: 404,
+          message: 'Not Found'
+        },
+        error: new ValidationError('Not Found')
+      }
+    }
+
     return await this.todoRepository.delete(id);
   }
 }
